refactoring and making random collectibles

+++ src/validation/validate_size.c
@@ -6,7 +6,7 @@
 /*   By: bmoreira <bmoreira@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/29 21:18:49 by bmoreira          #+#    #+#             */
-/*   Updated: 2025/11/04 19:39:47 by bmoreira         ###   ########.fr       */
+/*   Updated: 2025/11/07 15:08:21 by bmoreira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void	validate_size(t_map *map)
 {
-	size_t	row;
-	size_t	col;
+	int	row;
+	int	col;
 
 	row = 0;
 	col = ft_strlen(map->grid[row]);
 	while (map->grid[row])
-		if (ft_strlen(map->grid[row++]) != col)
+		if ((int) ft_strlen(map->grid[row++]) != col)
 			error_handler(map->grid, "Irregular map.", EXIT_FAILURE);
 	if (row < 4 || col < 4)
 		error_handler(map->grid, "Map too small.", EXIT_FAILURE);
